Update according to review
