+++ charms/garm/tests/unit/test_charm.py
@@ -28,13 +28,7 @@ from charm import (
     _proxy_environment,
     render_garm_toml,
 )
-from charm_state import SSHDebugInfo
-from garm_api import (
-    GarmApiError,
-    GarmConnectionError,
-    build_tmate_env_snippet,
-    prepend_after_shebang,
-)
+from garm_api import GarmApiError, GarmConnectionError
 from github_reconciler import DEFAULT_GITHUB_ENDPOINT
 from scaleset_reconciler import ScalesetSpec
 
@@ -529,32 +523,6 @@ def test_maybe_first_run_skips_on_missing_credential_key():
         GarmCharm._maybe_first_run(charm)
 
     mock_client_cls.assert_not_called()
-
-
-def test_ssh_debug_info_env_vars():
-    """
-    arrange: One SSHDebugInfo connection.
-    act: Call build_tmate_env_snippet() and prepend_after_shebang().
-    assert: The patched script starts with the shebang, then contains
-            all four TMATE_SERVER_* variables with correct values.
-    """
-    conn = SSHDebugInfo(
-        host="10.10.0.5",
-        port=2222,
-        rsa_fingerprint="SHA256:rsa",
-        ed25519_fingerprint="SHA256:ed",
-    )
-    base_script = "#!/bin/bash\necho 'hello'\n"
-
-    snippet = build_tmate_env_snippet([conn])
-    patched = prepend_after_shebang(base_script, snippet)
-
-    assert patched.startswith("#!/bin/bash\n")
-    assert "TMATE_SERVER_HOST=10.10.0.5" in patched
-    assert "TMATE_SERVER_PORT=2222" in patched
-    assert "TMATE_SERVER_RSA_FINGERPRINT=SHA256:rsa" in patched
-    assert "TMATE_SERVER_ED25519_FINGERPRINT=SHA256:ed" in patched
-    assert "echo 'hello'" in patched
 
 
 def test_proxy_environment_happy_path():
